refactor: rename profiles API file
